Add files via upload
Caesar Cipher:
The Caesar cipher is an encryption techniques. It works by shifting each letter of the plaintext by a certain number of positions down the alphabet. For example, if the shift value is 3, then each letter in the plaintext would be replaced by the letter three positions down the alphabet. So "A" becomes "D", "B" becomes "E", and so on. The receiver of the encrypted message can then shift each letter back up the alphabet by the same number of positions to reveal the original plaintext. The Caesar cipher is easy to implement but it is also very easy to break, especially if the attacker knows the encryption technique being used.

Brute-Force Attack on Caesar Cipher:
A brute-force attack on the Caesar cipher involves trying every possible shift value until the plaintext is found. Since there are only 25 possible shift values, this is easily done by a computer. The brute-force attack involves decrypting the ciphertext using each shift value, and then checking whether the resulting plaintext makes sense.