Add files via upload
Caesar: Implements the Caesar cipher, a simple substitution cipher that shifts each letter by a fixed amount (the "key"). This class allows for both encryption and decryption of plaintext.

Cipher: An interface that defines the basic methods for encryption and decryption of text. Any class that implements this interface must provide concrete implementations of these methods.

Substitution: An abstract class that implements the Cipher interface and provides concrete implementations of the encrypt and decrypt methods, based on abstract methods for encrypting and decrypting individual characters.

Playfair: Implements the Playfair cipher, a polygraphic substitution cipher that encrypts pairs of letters. This class generates a key matrix that is used for encryption and decryption.

Vigenere: Implements the Vigenere cipher, a type of substitution cipher that uses a keyword to determine the amount by which each letter is shifted. This class allows for both encryption and decryption of plaintext.